kpi_refresh_system/requirements.txt
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
schedule>=1.2.0
loguru>=0.7.0
click>=8.1.0
tenacity>=8.1.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
rich>=13.0.0
streamlit>=1.28.0
plotly>=5.17.0
pytz>=2023.3